+++ src/lib/settings/Gateway.js
@@ -262,6 +262,7 @@ class Gateway extends GatewayStorage {
 	 * @private
 	 */
 	async _ready() {
+		if (typeof this.client[this.type] === 'undefined') return null;
 		const promises = [];
 		const keys = await this.provider.getKeys(this.type);
 		for (let i = 0; i < keys.length; i++) {