+++ src/lib/settings/GatewayDriver.js
@@ -1,6 +1,5 @@
 const Gateway = require('./Gateway');
 const SettingResolver = require('../parsers/SettingResolver');
-const { Guild, User } = require('discord.js');
 
 /**
  * Gateway's driver to make new instances of it, with the purpose to handle different databases simultaneously.
@@ -55,10 +54,10 @@ class GatewayDriver {
 
 	/**
 	 * The data schema Klasa uses for guild configs.
-	 * @since 0.3.0
+	 * @since 0.5.0
 	 * @readonly
 	 */
-	get defaultDataSchema() {
+	get guildsSchema() {
 		return {
 			prefix: {
 				type: 'string',
@@ -89,6 +88,34 @@ class GatewayDriver {
 			},
 			disabledCommands: {
 				type: 'command',
+				default: [],
+				min: null,
+				max: null,
+				array: true,
+				configurable: true,
+				sql: 'TEXT'
+			}
+		};
+	}
+
+	/**
+	 * The data schema Klasa uses for client-wide configs.
+	 * @since 0.5.0
+	 * @readonly
+	 */
+	get clientStorageSchema() {
+		return {
+			userBlacklist: {
+				type: 'user',
+				default: [],
+				min: null,
+				max: null,
+				array: true,
+				configurable: true,
+				sql: 'TEXT'
+			},
+			guildBlacklist: {
+				type: 'guild',
 				default: [],
 				min: null,
 				max: null,
@@ -173,44 +200,6 @@ class GatewayDriver {
 		if (!provider) throw `This provider (${engine}) does not exist in your system.`;
 
 		return engine;
-	}
-
-	/**
-	 * The validator function Klasa uses for guild configs.
-	 * @since 0.5.0
-	 * @param {(Object|string)} guildResolvable The guild to validate.
-	 * @returns {KlasaGuild}
-	 * @private
-	 */
-	async validateGuild(guildResolvable) {
-		if (guildResolvable) {
-			let value;
-
-			if (typeof guildResolvable === 'string' && /^\d{17,19}$/.test(guildResolvable)) value = this.client.guilds.get(guildResolvable);
-			else if (guildResolvable instanceof Guild) value = guildResolvable;
-			if (value) return value;
-		}
-
-		throw new Error('The parameter <Guild> expects either a Guild ID or a Guild Instance.');
-	}
-
-	/**
-	 * The validator function Klasa uses for user configs.
-	 * @since 0.5.0
-	 * @param {(Object|string)} userResolvable The user to validate.
-	 * @returns {KlasaUser}
-	 * @private
-	 */
-	async validateUser(userResolvable) {
-		if (userResolvable) {
-			let value;
-
-			if (typeof userResolvable === 'string' && /^\d{17,19}$/.test(userResolvable)) value = await this.client.users.fetch(userResolvable);
-			else if (userResolvable instanceof User) value = userResolvable;
-			if (value) return value;
-		}
-
-		throw new Error('The parameter <User> expects either a User ID or a User Instance.');
 	}
 
 }